+++ gn2pg/main.py
@@ -4,16 +4,15 @@
 
 import argparse
 import logging
-import logging.config
 import sys
 from logging.handlers import TimedRotatingFileHandler
 
 from toml import TomlDecodeError
 
 from gn2pg import _, __version__, metadata
-from gn2pg.check_conf import Gn2PgConf, edit_config, list_configs, read_config
+from gn2pg.check_conf import Gn2PgConf
 from gn2pg.env import CONFDIR, LOGDIR
-from gn2pg.helpers import full_download, init, update
+from gn2pg.helpers import full_download, init, manage_configs, update
 from gn2pg.store_postgresql import PostgresqlUtils
 from gn2pg.utils import BColors
 
@@ -32,10 +31,15 @@ def arguments(args):
         :obj:`argparse.Namespace`: command line parameters namespace
     """
     # Get options
-
     parser = argparse.ArgumentParser(description="Gn2Pg Client app")
-    root_subparser = parser.add_subparsers(help="Global commands", required=False)
-    root_parser = root_subparser.add_parser("manage", help="Manage configs")
+
+    subparser = parser.add_subparsers(help="Config management commands", required=True)
+
+    config_parser = subparser.add_parser("config", help="Manage configs")
+    download_parser = subparser.add_parser("download", help="Manage downloads")
+    db_parser = subparser.add_parser("db", help="Manage downloads")
+
+    # Global commands
     parser.add_argument(
         "-V",
         "--version",
@@ -43,61 +47,78 @@ def arguments(args):
         action="version",
         version=f"%(prog)s v{__version__}",
     )
-    out_group = parser.add_mutually_exclusive_group()
-    out_group.add_argument(
+    output_group = parser.add_mutually_exclusive_group()
+    output_group.add_argument(
         "-v",
         "--verbose",
         help=_("Increase output verbosity"),
         action="store_true",
     )
-    out_group.add_argument("-q", "--quiet", help=_("Reduce output verbosity"), action="store_true")
-    parser.add_argument(
+    output_group.add_argument(
+        "-q", "--quiet", help=_("Reduce output verbosity"), action="store_true"
+    )
+
+    # Config commands
+    config_group = config_parser.add_mutually_exclusive_group(required=True)
+    config_group.add_argument(
         "--init",
-        help=_("Initialize the TOML configuration file"),
-        action="store_true",
+        nargs="?",
+        type=str,
+        const="config.toml",
+        help="Initialize the TOML configuration file",
     )
-    # parser.add_argument(
-    #     "--edit",
-    #     help=_("Edit the TOML configuration file in default editor"),
-    #     action="store_true",
-    # )
-    parser.add_argument(
-        "--json-tables-create",
-        help=_("Create or recreate json tables"),
-        action="store_true",
-    )
-    root_parser.add_argument(
-        "--list-configs",
+    config_group.add_argument(
+        "--list",
         help=_("List config files"),
         action="store_true",
     )
-    root_parser.add_argument(
-        "--read-config",
-        help=_("List config files"),
-        action="store_true",
+    config_group.add_argument(
+        "--read",
+        nargs="?",
+        type=str,
+        default=None,
+        const="empty",
+        help=_("Select and view config file"),
     )
-    root_parser.add_argument(
-        "--edit-config",
-        help=_("List, choose and edit a config file"),
-        action="store_true",
+    config_group.add_argument(
+        "--edit",
+        nargs="?",
+        type=str,
+        default=None,
+        const="empty",
+        help=_("Select and view config file"),
     )
-    customscript_group = parser.add_mutually_exclusive_group()
-    customscript_group.add_argument(
+
+    # database commands
+    db_group = db_parser.add_mutually_exclusive_group(required=True)
+    db_group.add_argument(
         "--custom-script",
         nargs="?",
+        type=str,
         help=_(
             '''Execute custom SQL script in DB, default is "to_gnsynthese".
         You can also use your own script by using absolute file path instead of "to_gnsynthese"'''
         ),
     )
-    download_group = parser.add_mutually_exclusive_group()
+    db_group.add_argument(
+        "--json-tables-create",
+        help=_("Create or recreate json tables"),
+        action="store_true",
+    )
+
+    # Download commands
+    download_group = download_parser.add_mutually_exclusive_group(required=True)
+
     download_group.add_argument("--full", help=_("Perform a full download"), action="store_true")
     download_group.add_argument(
         "--update",
         help=_("Perform an incremental download"),
         action="store_true",
     )
-    parser.add_argument("file", nargs="?", help="Configuration file name")
+
+    for p in (db_parser, download_parser):
+        p.add_argument("file", nargs="?", help="Configuration file name")
+
     return parser.parse_args(args)
 
 
@@ -107,6 +128,7 @@ def main(args) -> None:
     Args:
       args ([str]): command line parameter list
     """
+    logger.info("<main> begin")
     epilog = f"""\
 {sh_col.color('okblue')}{sh_col.color('bold')}{metadata.PROJECT} \
 {sh_col.color('endc')}{sh_col.color('endc')} \
@@ -121,30 +143,55 @@ def main(args) -> None:
     print(epilog)
 
     # Create $HOME/tmp directory if it does not exist
+
+    # Set up logging
+    setup_logging()
+    # Get command line arguments
+    args = arguments(args)
+    set_logging_level(args)
+
+    logger.info(_("%s, version %s"), sys.argv[0], __version__)
+    logger.debug("Args: %s", args)
+    logger.debug("Arguments: %s", sys.argv[1:])
+
+    if "config" in sys.argv:
+        handle_config_commands(args)
+
+    if any(cmd in ["download", "db"] for cmd in sys.argv):
+        if args.file is None:
+            logger.critical(_("You must provide a config file"))
+            sys.exit(0)
+        try:
+            cfg_ctrl = Gn2PgConf(args.file)
+        except TomlDecodeError:
+            logger.critical(_("Incorrect content in TOML configuration %s"), args.file)
+            sys.exit(0)
+
+        if "db" in sys.argv:
+            handle_database_commands(args, cfg_ctrl)
+        if "download" in sys.argv:
+            handle_download_commands(args, cfg_ctrl)
+
+
+def setup_logging() -> None:
+    """Set up logging configuration."""
     LOGDIR.mkdir(parents=True, exist_ok=True)
 
-    # create file handler which logs even debug messages
     filehandler = TimedRotatingFileHandler(
         str(LOGDIR / (__name__ + ".log")),
         when="midnight",
         interval=1,
         backupCount=100,
     )
-    # create console handler with a higher log level
-    # ch = logging.StreamHandler()
-    # create formatter and add it to the handlers
     formatter = logging.Formatter(
         "%(asctime)s - %(levelname)s - %(module)s:%(funcName)s - %(message)s"
     )
     filehandler.setFormatter(formatter)
-    # ch.setFormatter(formatter)
-    # add the handlers to the logger
     logger.addHandler(filehandler)
-    # logger.addHandler(ch)
 
-    # Get command line arguments
-    args = arguments(args)
-    # Define verbosity
+
+def set_logging_level(args) -> None:
+    """Set the logging level based on command line arguments."""
     if args.verbose:
         logger.setLevel(logging.DEBUG)
     elif args.quiet:
@@ -152,81 +199,65 @@ def main(args) -> None:
     else:
         logger.setLevel(logging.INFO)
 
-    logger.info(_("%s, version %s"), sys.argv[0], __version__)
-    logger.debug("Args: %s", args)
-    logger.info("Arguments: %s", sys.argv[1:])
 
-    if "manage" not in sys.argv:
+def handle_download_commands(args, cfg_ctrl) -> bool:
+    """Handle commands that are not related to 'manage'."""
 
-        if args.file is None:
-            logger.critical(_("You must provide a config file"))
-            return None
+    logger.info("<handle_data_commands>")
 
-        if not (CONFDIR / args.file).is_file():
-            logger.critical(_("Configuration file %s does not exist"), str(CONFDIR / args.file))
-            return None
+    if not (CONFDIR / args.file).is_file():
+        logger.critical(_("Configuration file %s does not exist"), str(CONFDIR / args.file))
+        return False
 
-        # If required, first create YAML file
-        if args.init:
-            logger.info(_("Creating TOML configuration file"))
-            init(args.file)
-            return None
+    logger.info(_("Getting configuration data from %s"), args.file)
 
-        logger.info(_("Getting configuration data from %s"), args.file)
-        try:
-            cfg_ctrl = Gn2PgConf(args.file)
-        except TomlDecodeError:
-            logger.critical(_("Incorrect content in TOML configuration %s"), args.file)
-            sys.exit(0)
-        cfg_source_list = cfg_ctrl.source_list
-        cfg = list(cfg_source_list.values())[0]
-        logger.info(
-            _("config file have {len(cfg_source_list)} source(s) wich are : %s"),
-            ", ".join(cfg_source_list.keys()),
-        )
+    if args.full:
+        logger.info(_("Perform full action"))
+        full_download(cfg_ctrl)
 
-        manage_pg = PostgresqlUtils(cfg)
+    if args.update:
+        logger.info(_("Perform update action"))
+        update(cfg_ctrl)
 
-        if args.json_tables_create:
-            logger.info(_("Create, if not exists, json tables"))
-            manage_pg.create_json_tables()
+    # Further processing...
+    return True
 
-        if args.custom_script:
-            logger.info(_("Execute custom script %s on db"), args.custom_script)
-            manage_pg.custom_script(args.custom_script)
 
-        if args.full:
-            logger.info(_("Perform full action"))
-            full_download(cfg_ctrl)
+def handle_database_commands(args, cfg_ctrl) -> None:
+    """Handle commands related to 'config'."""
 
-        if args.update:
-            logger.info(_("Perform update action"))
-            update(cfg_ctrl)
-    # Edit yaml config file
-    # if args.edit:
-    #     logger.info(_("Editing TOML configuration file"))
-    #     edit(args.file)
-    #     return None
-    if "manage" in sys.argv:
-        if args.list_configs:
-            logger.info(_("List config files"))
-            list_configs()
-            return None
+    cfg_source_list = cfg_ctrl.source_list
+    cfg = list(cfg_source_list.values())[0]
+    logger.info(
+        _("config file have {len(cfg_source_list)} source(s) wich are : %s"),
+        ", ".join(cfg_source_list.keys()),
+    )
 
-        if args.read_config:
-            logger.info(_("Read configs"))
-            read_config()
-            return None
+    manage_pg = PostgresqlUtils(cfg)
 
-        if args.edit_config:
-            logger.info(_("Read configs"))
-            edit_config()
-            return None
-        print(dir(args))
+    if args.json_tables_create:
+        logger.info(_("Create, if not exists, json tables"))
+        manage_pg.create_json_tables()
+    if args.custom_script:
+        logger.info(_("Execute custom script %s on db"), args.custom_script)
+        manage_pg.custom_script(args.custom_script)
 
-    # Get configuration from file
 
-    return None
+def handle_config_commands(args) -> None:
+    """Handle commands related to 'config'."""
+    print(args)
+    if args.init:
+        logger.info(_(f"Creating TOML configuration file {args.init}"))
+        init(args.init)
+    if args.list:
+        logger.info(_("List config files"))
+        manage_configs("list")
+    if args.read:
+        logger.info(_("Read configs"))
+        manage_configs("read", args.read)
+    if args.edit:
+        logger.info(_("Read configs"))
+        manage_configs("edit", args.edit)
 
 
 def run():